Refactor compute_eFC function to use torch.float16 for improved memory usage and performance

+++ pipelines/dataset_to_interhemispherical_eFC.py
@@ -33,7 +33,7 @@ def compute_eFC(edge_features):
     # Compute the correlation coefficient matrix in smaller chunks
     num_edges = edge_features.shape[0]
     chunk_size = 64  # Adjust chunk size based on available memory
-    eFC = torch.zeros((num_edges, num_edges), dtype=torch.float32)
+    eFC = torch.zeros((num_edges, num_edges), dtype=torch.float16)
 
     for i in range(0, num_edges, chunk_size):
         end_i = min(i + chunk_size, num_edges)
